Add functionality for response rendering.
Added functionality to render response when Accept HTTP header is */*,
text/html, application/json and application/xml. It injects the
content-type header in response as per the response type set by Accept
Header. ModelAndView struct is created having 3 fields:
1. Model
2. View
3. ContentType of response

This would be the response of every controller which is inspected using
reflection when we call the controller method as per routes.cfg
 #Todo Convert routes.cfg to yaml file for more modern style of
 properties file.
Should try to resolve #14.