Update server dependencies during deploy.
app deploy -n 'appName' -d 'dir' command does deploy. As a part of this it fetches all packages required that is gokul server and its dependencies. It issues the command "go get -d github.com/vipulbhale/gokul/server. However because of lack of -u flag it doesn't update the existing packages. This is causing issues while doing manual testing. However this flag (-u) needs to be removed later on as it may impact the stability of server dependencies stack in production.
This should resolve #15.