Fix the issue related to http 400/404 for app.
Try#1 to handle the issue related to the 400/404.
In the server handled the situation to send 400 when accept header is
missing.
In the server handled the situation to send 404 when requests are not in
router.
In the server handled the situation to send 400 when request's Accept
header doesn't match with response type.