Fix issue related to content negotiation.
Now added functionality to parse accept header using regex. Then sorted
it based on quality. Based on the topmost quality if content-type
corresponding to it existed and view's availability, then only corresponding  to it render the content. Else send http 406 for now. More functionality needs to be added more content-negotiation.